moved entities to model package
