--- src/main/java/app/entities/Receipt.java
@@ -1,4 +1,0 @@
-package app.entities;
-
-public class Receipt {
-}
