+++ src/main/java/app/model/entities/Item.java
@@ -1,4 +1,4 @@
-package app.entities;
+package app.model.entities;
 
 public record Item(double price_pr_unit, double length, String unit, String description) {
 }
