+++ src/main/java/app/model/entities/ItemList.java
@@ -1,4 +1,4 @@
-package app.entities;
+package app.model.entities;
 
 import java.util.ArrayList;
 import java.util.List;
