+++ src/main/java/app/model/entities/Receipt.java
@@ -1,0 +1,4 @@
+package app.model.entities;
+
+public class Receipt {
+}
