+++ src/main/java/app/model/entities/Shed.java
@@ -1,4 +1,4 @@
-package app.entities;
+package app.model.entities;
 
 public class Shed {
