+++ src/main/java/app/persistence/UserMapper.java
@@ -8,6 +8,9 @@ import java.util.ArrayList;
 import java.util.List;
 
 import app.entities.*;
+import app.model.entities.Customer;
+import app.model.entities.Salesperson;
+import app.model.entities.User;
 
 
 public class UserMapper {
