+++ src/test/java/app/persistence/UserMapperTest.java
@@ -1,12 +1,10 @@
 package app.persistence;
 
-import app.entities.Customer;
-import app.entities.User;
+import app.model.entities.Customer;
+import app.model.entities.User;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class UserMapperTest {
     int id;
